Update advanced-usage.rst
Just a few changes to address run-on sentences, punctuation and spelling.

Unresolved issues. If this is too pedantic, just let me know...

This document mixes whether 'SSH-MITM proxy server' is a name by itself or not. For example, consider this sentence:

"SSH-MITM proxy server is capable of advanced man-in-the-middle attacks."
This is fine, treating 'SSH-MITM proxy server' as a name or term-of-the-art.
Consider this sentence:
"If agent forwarding is not possible, the SSH-MITM proxy server can accept the
public key authentication request and redirect the session to a honey pot."
Also fine, but now its descriptive rather than a name.
Finally, in some places we have just SSH-MITM as a name, which of course it is. But since it's 'SSH-MITM proxy server' everywhere else, we should probably add proxy server here or remove it almost everywhere else....

My suggestion would be to describe it as 'The SSH-MITM proxy server' in the first sentence, and then just call it 'SSH-MITM' from there on.

Again, possibly too pedantic....