Update images?
Just my two cents:
The images are confusing. Of the three images, the only one where I would call the laptop `client` is the one where the command is `ssh remote`. In the other ones, where ssh acts as a port forwarder, I could call it `proxy` or `port forwarder` or something - but not `client`. The client is actually missing in those images.
If you disagree, we can just leave it as is. Otherwise I will take a crack at modifying the images.